// src/app/modules/main-container/models/conversation-record.ts
export class ChatRowRecord {
  title: string;
  firstUser: string;
  secondUser: string;
  username: string;
  fullname: string;
  profilePictureId: string;
  groupId: string;
  type: string;
  regDateTime: string;
  content: string;
  isDeletedChat: boolean;
  isSeen: boolean;
  attachments: string;
  regUser: string;
  unReadMessages: string;
  lastMessageId: number;
  owner: string;
  timestamp: string;
  firstMessageId: string;
  receiverId: string;
  typeTitle: string;
  unreadCount: number = 0;
  contentId: number = 0;
  id: number = 0;
}